bug fixing

+++ commands/roll.js
@@ -22,8 +22,8 @@ module.exports = {
                 if (isNaN(parsedRepeat)) {
                     throw 'repeat argument must be a number!'
                 }
-                else if (parsedRepeat < 1 || parsedRepeat > 100) {
-                    throw 'repeat argument must be a number between 1 and 100!'
+                else if (parsedRepeat < 1 || parsedRepeat > 20) {
+                    throw 'repeat argument must be a number between 1 and 20!'
                 }
                 else if (parsedRepeat === 1) {
                     //pass
@@ -31,13 +31,22 @@ module.exports = {
                 else {
                     const embed = new MessageEmbed().setTitle(`Wynik rzutu - (${dices})`).setDescription(`liczba powtórzeń - ${repeat}`)
                     let sum = 0
+
                     for (let i = 0; i < parsedRepeat; i++) {
                         const roll = new rpgDiceRoller.DiceRoll(dices)
-
-                        embed.addField(`Rzut #${i + 1}`, `\`\`\` 🎲 ${roll}\`\`\``)
+                        let rollResult = `\`\`\` 🎲 ${roll}\`\`\``
+                        if (rollResult.length > 1000) {
+                            throw 'roll result is too long for discord message standards.'
+                        }
+                        embed.addField(`Rzut #${i + 1}`, rollResult)
                         sum += roll.total
                     }
-                    embed.addField('Suma', sum)
+
+                    embed.addField('Wynik', sum)
+
+                    if (embed.length > 6000) {
+                        throw 'roll result is too long for discord message standards.'
+                    }
 
                     if (message) {
                         message.delete()
@@ -46,21 +55,29 @@ module.exports = {
                     return embed
                 }
             }
+
             const roll = new rpgDiceRoller.DiceRoll(dices)
             const embed = new MessageEmbed()
             let rolls = `\`\`\` 🎲 ${roll.rolls.join(' ')}\`\`\``
+
             if (rolls.length > 2048) {
                 rolls = rolls.slice(0, 2030) + '...]```'
             }
+
             embed
                 .setTitle(`Wynik rzutu - (${roll.notation})`)
                 .setDescription(rolls)
-                .addField('suma', roll.total)
-            console.log(roll.averageTotal)
+                .addField('Wynik', roll.total)
+
             if ((roll.total === roll.maxTotal) && (roll.maxTotal !== roll.minTotal) && roll.averageTotal > 5) {
                 embed.setImage('https://media.giphy.com/media/9y2rmR2dv7rLq/giphy.gif')
                 embed.setColor('#00ff00')
             }
+            else if ((roll.total === roll.minTotal) && (roll.maxTotal !== roll.minTotal) && roll.averageTotal > 5) {
+                embed.setImage('https://media1.tenor.com/images/8474b0f567704b752354d6e5a589784b/tenor.gif?itemid=6118887')
+                embed.setColor('#ff0000')
+            }
+
             if (message) {
                 message.delete()
                 message.reply('', { embed })
@@ -71,6 +88,7 @@ module.exports = {
         catch (e) {
             var embed = new MessageEmbed()
                 .setTitle('Wynik rzutu - Błąd')
+                .setColor('#ffff00')
                 .setDescription(e)
 
             if (message) {
